Refactor airship model structure and update trajectory generation functionality
- Introduced a new observer class for disturbance estimation in the airship model.
- Consolidated airship-related modules into a single AirshipModel directory for better organization.
- Updated trajectory generation methods to enhance clarity and maintainability.
- Removed deprecated files and cleaned up the project structure for improved navigation.
