+++ AirshipModel/rotation_matrices.py
@@ -1,5 +1,5 @@
 """
-utils.py
+rotation_matrices.py
 refer to Nonlinear adaptive trajectory tracking control for a stratospheric airship with parametric uncertainty
     Error-constrained fixed-time trajectory tracking control for a stratospheric airship with disturbances
 """
